Fix typo in save

+++ scrapers/scrape_functions.py
@@ -532,14 +532,15 @@ def remove_deleted_offerings(timetable_globals):
 		alt_course = Course.query.filter(Course.number == offering_course.number, Course.department_id == offering_course.department_id, Course.name != offering_course.name, Course.name.contains(offering_course.name)).first()
 		if alt_course:
 			updated_offering = Offering.query.filter(Offering.course_id == alt_course.id, Offering.term_id == offering.term_id).first()
-			for user in all_users:
-				user.drop(offering)
-				user.take(updated_offering)
+			if updated_offering:
+				for user in all_users:
+					user.drop(offering)
+					user.take(updated_offering)
 
-			print_alert("UPDATED NAME: " + repr(offering_course.name) + " to " + repr(alt_course.name))
-			db.session.delete(offering)
-			db.session.commit()
-			continue
+				print_alert("UPDATED NAME: " + repr(offering_course.name) + " to " + repr(alt_course.name))
+				db.session.delete(offering)
+				db.session.commit()
+				continue
 
 		# Ignore if ORC data from the higher-priority timetable
 		if offering.get_term() is not None:
